Restrict version of a dependency
Lock in a version of a dependency, as directed in Bundler's docs.

https://bundler.io/gemfile.html

In this example, we'll accept any version at least as recent as 4.31.0,
but nothing 5.0.0 and beyond. This may be necessary because application
code is incompatible with newer versions, but we trust the gem's
developers to maintain compatibility within 4.x versions.